agent-engines/agent-starter-pack/pyproject.toml
[project]
name = "gcp-agent-starter"
version = "0.1.0"
description = "CLI tool to create GCP-based AI agent projects from templates"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.7",
    "cookiecutter>=2.5.0", 
    "google-cloud-aiplatform>=1.38.1",
    "inquirer>=3.1.3",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
gcp-agent-starter = "src.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build]
artifacts = ["src"]
